refactor(tags): reorganiza a layout ui/model/api/hooks como notes
