--- src/actions/tags/index.ts
@@ -1,2 +1,0 @@
-export * from './getActiveTagGroup'
-export * from './setActiveTagGroup'
