+++ src/features/tags/hooks/useAvailableTags.tsx
@@ -1,7 +1,7 @@
 import { useTagStore } from '../state/store'
 import { AvailableTags, TagGroup } from '../model/tags'
 import { useTranslation } from 'react-i18next'
-import { translateTagGroup } from '../utils/translateTagGroup'
+import { translateTagGroup } from '../model/translateTagGroup'
 
 export const useAvailableTags = (): AvailableTags => {
 	const availableTags = useTagStore((state) => state.availableTags)
