+++ src/features/tags/index.ts
@@ -2,12 +2,12 @@ export type { Tag, TagGroup, AvailableTags } from './model/tags'
 export { emptyTagGroup, defaultAvialableTags, getTagGroup } from './model/tags'
 export { eisenhowerTagGroup, devTagGroup } from './model/defaultTags'
 export { getHighestPriority } from './model/getHighestPriority'
-export { translateTagGroup } from './utils/translateTagGroup'
-export { fetchTags, saveTags } from './repository'
+export { translateTagGroup } from './model/translateTagGroup'
+export { fetchTags, saveTags } from './api/repository/tagRepositoryFactory'
 export { useTagStore } from './state/store'
 export { useTagsQuery } from './hooks/useTagsQuery'
 export { useActualTagGroup } from './hooks/useActualTagGroup'
 export { useAvailableTags } from './hooks/useAvailableTags'
 export { useUserSelectedTags } from './hooks/useUserSelectedTags'
-export { EnableTags } from './components/EnableTags'
-export { default as TagGroupSelect } from './components/TagGroupSelect'
+export { EnableTags } from './ui/EnableTags'
+export { default as TagGroupSelect } from './ui/TagGroupSelect'
