+++ src/modules/TaskBoard/components/taskList/components/AddNewTaskInput.tsx
@@ -4,11 +4,11 @@ import { getNewTask, isThisTaskDescriptionValid } from '@/modules/TaskBoard/mode
 import { toast } from 'sonner'
 import getErrorMessageForTheUser from '@/shared/lib/getErrorMessageForTheUser'
 import { KeyboardEvent, useState } from 'react'
-import { useTagStore } from '@/features/tags/state/store'
+import { useTagStore } from '@/features/tags'
 import { useTranslation } from 'react-i18next'
 import { TeaxtareaWithActions } from '@/shared/ui/molecules/TextAreaWithActions'
-import TagGroupSelect from '@/features/tags/components/TagGroupSelect'
-import { useUserSelectedTags } from '@/features/tags/hooks/useUserSelectedTags'
+import { TagGroupSelect } from '@/features/tags'
+import { useUserSelectedTags } from '@/features/tags'
 import { addTaskInFirstColumn } from '../useCase/addTask'
 import { useTaskBoardQuery } from '@/modules/TaskBoard/hooks/useTaskBoardQuery'
 import { sortListOfTasksInColumnsByPriority } from '../models/sortListOfTasksInColumnsByPriority'
